// canvas/Canvas.enums.ts
export enum NODE_TYPES {
  DEFAULT = 'DEFAULT',
  IF_ELSE = 'IF_ELSE',
  FOR_EACH = 'FOR_EACH',
  DO_IN_PARALLEL = 'DO_IN_PARALLEL',
  END = 'END',
  WAIT_FOR = 'WAIT_FOR',
}

export enum UTILITY_NODES {
  DO_IN_PARALLEL = 'DO_IN_PARALLEL',
  ONCE_ALL_COMPLETED = 'ONCE_ALL_COMPLETED',
  END = 'END',
}

export enum PLACEHOLDER_NODES {
  EMPTY_STATE_PLACEHOLDER = 'EMPTY_STATE_PLACEHOLDER',
  DROP_TARGET_PLACEHOLDER = 'DROP_TARGET_PLACEHOLDER',
}

export enum BLOCK_TYPES {
  IF_ELSE = 'IF_ELSE',
  FOR_EACH = 'FOR_EACH',
  DO_IN_PARALLEL = 'DO_IN_PARALLEL',
}

export enum DROP_DIRECTION {
  TOP = 'TOP',
  BOTTOM = 'BOTTOM',
  LEFT = 'LEFT',
  RIGHT = 'RIGHT',
}